backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg[binary,pool]==3.2.13
sqlalchemy[asyncio]==2.0.36
pgvector==0.3.6
google-cloud-storage==2.19.0
python-multipart==0.0.9
python-dotenv==1.0.1
google-cloud-speech==2.26.0
tenacity==8.2.3
google-cloud-aiplatform>=1.60.0
nltk==3.8.1
tiktoken==0.5.2
pypdf==3.17.4
python-dateutil==2.8.2
google-cloud-run==0.10.5
asyncpg==0.29.0
google-cloud-tasks>=2.16.3